sql-agent-google-sheet/requirements.txt
# Core AI/ML libraries
langchain>=0.1.0
langchain-community>=0.0.12
langchain-openai>=0.0.5
langchain-mcp-adapters>0.1.0
langgraph>=0.0.20

# MCP Integration
mcp>=1.0.0

# Data processing
pandas>=2.0.0
chardet>=5.0.0

# Web framework
streamlit>=1.30.0

# Configuration and utilities
python-dotenv>=1.0.0
typing-extensions>=4.0.0
pydantic>=2.0.0

# Additional dependencies
aiofiles>=23.0.0
uvicorn>=0.24.0
fastapi>=0.104.0
requests>=2.31.0
numpy>=1.24.0
sqlalchemy>=2.0.0

# Optional dev/testing dependencies (install with: pip install -r requirements.txt -r requirements-dev.txt)
# pytest>=7.0.0
# pytest-asyncio>=0.21.0
